+++ apps/wallet-mobile/src/features/Portfolio2/useCases/PortfolioDashboard/BalanceCard/BalanceCardSkeleton.tsx
@@ -2,19 +2,11 @@ import {useTheme} from '@yoroi/theme'
 import * as React from 'react'
 import SkeletonPlaceholder from 'react-native-skeleton-placeholder'
 
-type Props = {
-  cardHeight?: number
-}
-export const BalanceCardSkeleton = ({cardHeight = 84}: Props) => {
-  const {color} = useStyles()
+export const BalanceCardSkeleton = () => {
+  const {color} = useTheme()
   return (
     <SkeletonPlaceholder borderRadius={9} backgroundColor={color.gray_c100}>
-      <SkeletonPlaceholder.Item width="100%" height={cardHeight} />
+      <SkeletonPlaceholder.Item width="100%" height={122} />
     </SkeletonPlaceholder>
   )
-}
-
-const useStyles = () => {
-  const {color} = useTheme()
-  return {color} as const
 }
