+++ apps/wallet-mobile/src/features/Portfolio2/useCases/PortfolioTokensList/PortfolioDAppsTokenList/DAppTokenItem/DAppTokenItemSkeleton.tsx
@@ -2,7 +2,7 @@ import {useTheme} from '@yoroi/theme'
 import * as React from 'react'
 import SkeletonPlaceholder from 'react-native-skeleton-placeholder'
 
-export const DAppTokenITemSkeleton = () => {
+export const DAppTokenItemSkeleton = () => {
   const {color} = useTheme()
   return (
     <SkeletonPlaceholder backgroundColor={color.gray_c100}>
