+++ apps/wallet-mobile/src/features/Portfolio2/useCases/PortfolioTokensList/PortfolioWalletTokenList/PortfolioWalletTokenList.tsx
@@ -4,8 +4,8 @@ import * as React from 'react'
 import {FlatList, StyleSheet, Text, View} from 'react-native'
 
 import {Spacer} from '../../../../../components'
-import {useSearch} from '../../../../../Search/SearchContext'
-import {makeList} from '../../../../../utils'
+import {useSearch} from '../../../../../features/Search/SearchContext'
+import {makeList} from '../../../../../kernel/utils'
 import {Line} from '../../../common/Line'
 import {TokenEmptyList} from '../../../common/TokenEmptyList'
 import {useGetPortfolioBalance} from '../../../common/useGetPortfolioBalance'
