+++ apps/wallet-mobile/src/features/Send/useCases/ListAmountsToSend/AddToken/SelectTokenFromListScreen.tsx
@@ -8,12 +8,12 @@ import React from 'react'
 import {Alert, StyleSheet, TouchableOpacity, View} from 'react-native'
 
 import {Spacer, Text} from '../../../../../components'
+import {useSearch, useSearchOnNavBar} from '../../../../../features/Search/SearchContext'
 import {useMetrics} from '../../../../../kernel/metrics/metricsManager'
 import {TxHistoryRouteNavigation} from '../../../../../kernel/navigation'
 import {usePortfolioBalances} from '../../../../Portfolio/common/hooks/usePortfolioBalances'
 import {MediaGallery} from '../../../../Portfolio/common/MediaGallery/MediaGallery'
 import {TokenAmountItem} from '../../../../Portfolio/common/TokenAmountItem/TokenAmountItem'
-import {useSearch, useSearchOnNavBar} from '../../../../Search/SearchContext'
 import {useSelectedWallet} from '../../../../WalletManager/context/SelectedWalletContext'
 import {limitOfSecondaryAmountsPerTx} from '../../../common/constants'
 import {useOverridePreviousSendTxRoute} from '../../../common/navigation'
